analytics: extract context env info getter methods to util/analytics.ts for reusability.
